backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
mongoengine==0.27.0
python-dotenv==1.0.0
marshmallow==3.20.1
marshmallow-mongoengine==0.30.0
pymongo==4.5.0
requests==2.31.0
anthropic==0.40.0
ipdb
